Typescript graphql (#165)
* graphql schema split

* test

* load files

* gql types split using esm import/export

* remove dependencies

* package.json

* added new for NonNull and List types

* graphql

* moved readme

* changed soms spacing

* added more types

* Fixed Bug with Wrong Instructor for Common Names (#156)

* fixed bug using wrong inst array

* fixed the second usage

* Update schedule query to return courseOffering (#161)

Before it returned a courseType. This doesn't really make sense though since the client cares more about the courseOffering information. The user can still get the course data from courseOffering if they want it.

* Fixed instructor resolve returning null in offering (#163)

* Fixed instructor resolve returning null in offering

* Removed unnecessary assignment

* made connection param optional

* fixed some types that were redundant

* updated compiler options

* fixed type for null averageGPA
