+++ README.md
@@ -5,7 +5,9 @@ PeterPortal Public API provides software developers with easy-access to UC Irvin
 🔨 Built with:
 
 * Express.js
+* TypeScript
 * GraphQL
+* Serverless
 
 👯‍♂️ Works with:
