+++ helpers/grades.helper.ts
@@ -108,13 +108,33 @@ export function parseGradesParamsToSQL(query) : string{
 }
 
 export function fetchGrades(where: string) : GradeRawData {
-    let sqlStatement = "SELECT * FROM gradeDistribution";
-    return queryDatabase(null, where !== null ? sqlStatement + where : sqlStatement).all();
+    let sqlStatement = `SELECT 
+    year, 
+    quarter, 
+    department, 
+    department_name,
+    number, 
+    number_int,
+    code, 
+    section, 
+    title, 
+    instructor, 
+    type, 
+    gradeACount, 
+    gradeBCount,
+    gradeCCount,
+    gradeDCount,
+    gradeFCount,
+    gradePCount,
+    gradeNPCount,
+    gradeWCount,
+    NULLIF(averageGPA, '') as averageGPA FROM gradeDistribution`;
+    return queryDatabase(where !== null ? sqlStatement + where : sqlStatement).all();
 }
 
 export function fetchInstructors(where: string) : string[] {
     let sqlStatement = "SELECT DISTINCT instructor FROM gradeDistribution";
-    return queryDatabase(null, where !== null ? sqlStatement + where : sqlStatement).all().map(result => result.instructor);
+    return queryDatabase(where !== null ? sqlStatement + where : sqlStatement).all().map(result => result.instructor);
 }
 
 //For GraphQL API
@@ -131,10 +151,10 @@ export function fetchAggregatedGrades(where: string) : GradeDist {
     AVG(NULLIF(averageGPA, '')) as average_gpa,
     COUNT() as count FROM gradeDistribution`;
     
-    return queryDatabase(null, where != null ? sqlStatement + where : sqlStatement).get();
+    return queryDatabase(where != null ? sqlStatement + where : sqlStatement).get();
 }
 
-function queryDatabase(connection: Database, statement: string) {
+function queryDatabase(statement: string, connection?: Database) {
     if (!connection) {
         connection = new db(path.join(__dirname, '../db/db.sqlite'));
     }
@@ -173,8 +193,8 @@ export function fetchCalculatedData(where) : GradeCalculatedData{
     instructor,
     type FROM gradeDistribution`;
 
-    result.gradeDistribution = queryDatabase(connection, where !== null ? sqlFunction + where : sqlFunction).get();
-    result.courseList = queryDatabase(connection, where !== null ? sqlCourseList + where : sqlCourseList).all();
+    result.gradeDistribution = queryDatabase(where !== null ? sqlFunction + where : sqlFunction, connection).get();
+    result.courseList = queryDatabase(where !== null ? sqlCourseList + where : sqlCourseList, connection).all();
     connection.close();
     return result;
 }
