+++ tests/int/graphql/schedule.graphql.test.ts
@@ -9,35 +9,31 @@ describe('POST /graphql/', () => {
     .post('/graphql/')
     .send({query:`{
         schedule(year: 2019, quarter: "Fall", department:"COMPSCI", course_number: "161") {
-          title
-          id
-          offerings {
-            year
-            quarter
-            instructors {
-              ucinetid
-              name
-              department
-            }
-            final_exam
-            max_capacity
-            meetings {
-              time
-              building
-              days
-            }
-            num_section_enrolled
-            num_total_enrolled
-            num_on_waitlist
-            num_requested
-            num_new_only_reserved
-            units
-            restrictions
-            status
-            course {
-              title
-              id
-            }
+          year
+          quarter
+          instructors {
+            ucinetid
+            name
+            department
+          }
+          final_exam
+          max_capacity
+          meetings {
+            time
+            building
+            days
+          }
+          num_section_enrolled
+          num_total_enrolled
+          num_on_waitlist
+          num_requested
+          num_new_only_reserved
+          units
+          restrictions
+          status
+          course {
+            title
+            id
           }
         }
       }`})
@@ -48,10 +44,9 @@ describe('POST /graphql/', () => {
         expect(response.body).toHaveProperty('data');
         expect(response.body["data"]).toHaveProperty('schedule');
         expect(Array.isArray(response.body["data"]["schedule"])).toBeTruthy();
-        expect(response.body["data"]["schedule"][0]["title"]).toEqual("Design and Analysis of Algorithms");
-        expect(response.body["data"]["schedule"][0]["id"]).toEqual("COMPSCI161");
-        expect(Array.isArray(response.body["data"]["schedule"][0]["offerings"])).toBeTruthy();
-        expect(response.body["data"]["schedule"][0]["offerings"][0]).toEqual(
+        expect(response.body["data"]["schedule"][0]["course"]["title"]).toEqual("Design and Analysis of Algorithms");
+        expect(response.body["data"]["schedule"][0]["course"]["id"]).toEqual("COMPSCI161");
+        expect(response.body["data"]["schedule"][0]).toEqual(
           expect.objectContaining({
             "year": "2019",
             "quarter": "Fall",
@@ -78,13 +73,11 @@ describe('POST /graphql/', () => {
     .post('/graphql/')
     .send({query:`{
         schedule(year: 2019, quarter: "Fall", department:"COMPSCI", course_number: "100") {
-            offerings {
-                year
-                max_capacity
-                num_total_enrolled
-            }
+          year
+          max_capacity
+          num_total_enrolled
         }
-        }`})
+      }`})
     .set('Accept', 'application/json')
     .expect('Content-Type', /json/)
     .expect(200)
@@ -100,11 +93,9 @@ describe('POST /graphql/', () => {
   .post('/graphql/')
   .send({query:`{
       schedule(year: 2022, quarter: "Spring", instructor: "SMITH, J.") {
-          offerings {
-            instructors {
-              name
-            }
-          }
+        instructors {
+          name
+        }
       }
       }`})
   .set('Accept', 'application/json')
@@ -118,29 +109,17 @@ describe('POST /graphql/', () => {
       //check that all instructors are present.
       expect(response.body["data"]["schedule"]).toContainEqual(
         expect.objectContaining({
-          "offerings": [
-            {
-              "instructors": [ { "name": "Jaymi Lee Smith" } ]
-            }
-          ]
+            "instructors": [ { "name": "Jaymi Lee Smith" } ]
         })
       );
       expect(response.body["data"]["schedule"]).toContainEqual(
         expect.objectContaining({
-          "offerings": [
-            {
               "instructors": [ { "name": "John H Smith" } ]
-            }
-          ]
         })
       );
       expect(response.body["data"]["schedule"]).toContainEqual(
         expect.objectContaining({
-          "offerings": [
-            {
-              "instructors": [ { "name": "James N Smith" } ]
-            }
-          ]
+          "instructors": [ { "name": "James N Smith" } ]
         })
       );
   }));
@@ -151,11 +130,9 @@ describe('POST /graphql/', () => {
     .post('/graphql/')
     .send({query:`{
         schedule(year: 2019, quarter: 2019, department:"COMPSCI", course_number: "161") {
-            offerings {
-                year
-                max_capacity
-                num_total_enrolled
-            }
+            year
+            max_capacity
+            num_total_enrolled
         }
     }`})
     .set('Accept', 'application/json')
@@ -171,15 +148,32 @@ describe('POST /graphql/', () => {
 
 
 describe('POST /graphql/', () => {
-  it('GraphQL: Schedule query with invalid argument',  () => request
+  it('GraphQL: Schedule query with invalid quarter argument',  () => request
   .post('/graphql/')
   .send({query:`{
       schedule(year: 2019, quarter: 2019, department:"COMPSCI", course_number: "161") {
-          offerings {
-              year
-              max_capacity
-              num_total_enrolled
-          }
+          year
+          max_capacity
+          num_total_enrolled
+      }
+  }`})
+  .set('Accept', 'application/json')
+  .expect('Content-Type', /json/)
+  .expect(400)
+  .then((response) => {
+      expect(response.body).toHaveProperty('errors');
+      expect(response.body["errors"][0]["message"]).toMatch(/(2019)/i);
+  }));
+});
+
+describe('POST /graphql/', () => {
+  it('GraphQL: Schedule query without year',  () => request
+  .post('/graphql/')
+  .send({query:`{
+      schedule(quarter: 2019, department:"COMPSCI", course_number: "161") {
+          year
+          max_capacity
+          num_total_enrolled
       }
   }`})
   .set('Accept', 'application/json')
