+++ tests/unit/schedule.helper.unit.test.ts
@@ -1,5 +1,5 @@
 import { getCourseSchedules } from "../../helpers/schedule.helper";
-import { CourseGQL } from "../../types/websoc.types";
+import { CourseOffering } from "../../types/types";
 
 describe('Get Schedule of Classes', () => {
     it ('getCourseSchedules should return an array of results', async () => {
@@ -8,22 +8,21 @@ describe('Get Schedule of Classes', () => {
             department:"COMPSCI",
             courseNumber: "161"
         }
-        const schedules : CourseGQL[] = await getCourseSchedules(query);
+        const schedules : CourseOffering[] = await getCourseSchedules(query);
         expect(schedules).not.toBeNull() 
         expect(typeof schedules).toBe("object");
         expect(Array.isArray(schedules)).toBe(true);
         expect(schedules.length).toBeGreaterThan(0);
-        const course : CourseGQL = schedules[0];
+        const offering : CourseOffering = schedules[0];
         expect(schedules).toContainEqual(
-            expect.objectContaining({
-                "id": "COMPSCI161"
-            })
-        );
-        expect(course.offerings).toContainEqual(
             expect.objectContaining({
                 "year": "2019", 
                 "quarter": "Fall",
-                "course": "COMPSCI161"
+            })
+        );
+        expect(offering.course).toEqual(
+            expect.objectContaining({
+                "id": "COMPSCI161"
             })
         );
     });
