+++ types/websoc.types.ts
@@ -2,11 +2,13 @@
 export interface WebsocResponse {
     schools: School[]
 } 
+
 export interface School {
     schoolName: string;
     schoolComment: string;
     departments: Department[];
 }
+
 export interface Department {
     deptName: string;
     deptCode: string;
@@ -15,6 +17,7 @@ export interface Department {
     sectionCodeRangeComments: string[];
     courseNumberRangeComments: string[];
 }
+
 export interface Course {
     courseNumber: string;
     courseTitle: string;
@@ -22,6 +25,7 @@ export interface Course {
     prerequisiteLink: string;
     sections: Section[];
 }
+
 export interface Section {
     sectionCode: string;
     sectionType: string;
@@ -39,64 +43,18 @@ export interface Section {
     status: string;
     sectionComment: string;
 }
+
 export interface Meeting {
     days: string;
     time: string;
     bldg: string;
 }
+
 export interface EnrollmentCount {
     totalEnrolled: string;
     sectionEnrolled: string;
 }
 
-
-
-export interface CourseGQL {
-    id: string;
-    department: string;
-    number: string;
-    school: string;
-    title: string;
-    course_level: string;
-    department_alias: string[];
-    units: number[];
-    description: string;
-    department_name: string;
-    professor_history: string[];
-    prerequisite_tree: string;
-    prerequisite_list: string[];
-    prerequisite_text: string;
-    prerequisite_for: string[];
-    repeatability: string;
-    concurrent: string;
-    same_as: string;
-    restriction: string;
-    overlap: string;
-    corequisite: string;
-    ge_list: string[];
-    ge_text: string;
-    terms: string[];
-    offerings: CourseOffering[];
-}
-
-export interface CourseOffering {
-    year: string;
-    quarter: string;
-    instructors: string[];
-    final_exam: string;
-    max_capacity: number;
-    meetings: Meeting[];
-    num_section_enrolled: number;
-    num_total_enrolled: number;
-    num_new_only_reserved: number;
-    num_on_waitlist: number;
-    num_requested: number;
-    restrictions: string;
-    section: SectionGQL;
-    status: string;
-    units: number;
-    course: CourseGQL;
-}
 
 export interface SectionGQL {
     code: string;
